+++ .changeset/swift-cooks-drum.md
@@ -1,0 +1,5 @@
+---
+'@primer/react-brand': patch
+---
+
+Added experimental TextRevealAnimation component
