+++ packages/react/src/Testimonial/Testimonial.tsx
@@ -111,12 +111,9 @@ const Root = forwardRef(_Root)
  * Testimonial quote child element
  * <Testimonial.Quote>
  */
-type QuoteProps = {
-  children: string | Array<string | React.ReactElement<HTMLElement>>
-} & React.HTMLAttributes<HTMLElement> &
-  BaseProps<HTMLElement>
+type QuoteProps = React.HTMLAttributes<HTMLElement> & BaseProps<HTMLElement>
 
-function _Quote({children, className}: QuoteProps, ref) {
+function _Quote({children, className}: PropsWithChildren<QuoteProps>, ref) {
   const childrenArray = useMemo(() => React.Children.toArray(children), [children])
 
   // TODO: when Firefox supports :has() selector, we should use that instead of JS
