+++ packages/react/src/animation/TextRevealAnimation/TextRevealAnimation.module.css
@@ -1,0 +1,37 @@
+@keyframes wipe {
+  from {
+    transform: scaleX(1);
+  }
+  to {
+    transform: scaleX(0);
+  }
+}
+
+.TextRevealAnimation {
+  visibility: hidden;
+}
+
+.TextRevealAnimation--visible {
+  visibility: visible;
+}
+
+.TextRevealAnimation,
+.TextRevealAnimation--animated {
+  width: 100%;
+  min-height: 1lh;
+  position: relative;
+}
+
+.TextRevealAnimation--animated::after {
+  content: '';
+  display: inline;
+  position: absolute;
+  top: 0;
+  right: 0;
+  width: 100%;
+  height: 100%;
+  background-color: var(--brand-color-accent-primary);
+  transform-origin: right;
+  animation: wipe var(--brand-animation-duration-extended) forwards var(--brand-animation-easing-default);
+  animation-delay: var(--animation-delay);
+}
