+++ packages/react/src/animation/TextRevealAnimation/TextRevealAnimation.module.css.d.ts
@@ -1,0 +1,8 @@
+declare const styles: {
+  readonly "TextRevealAnimation": string;
+  readonly "TextRevealAnimation--visible": string;
+  readonly "TextRevealAnimation--animated": string;
+  readonly "wipe": string;
+};
+export = styles;
+
