+++ packages/react/src/animation/TextRevealAnimation/index.ts
@@ -1,0 +1,1 @@
+export * from './TextRevealAnimation'
