+++ packages/react/src/animation/index.ts
@@ -1,3 +1,4 @@
 export * from './AnimationProvider'
 export * from './Animate'
 export * from './useAnimation'
+export * from './TextRevealAnimation'
